CreateBookingRequest & GetCalendarRequest: validation attributes updated.

+++ VacationRental.Application/Calendar/GetCalendarRequest.cs
@@ -5,7 +5,7 @@ namespace VacationRental.Application.Calendar
 {
     public sealed class GetCalendarRequest
     {
-        [Range(1, Double.MaxValue)]
+        [Range(1, int.MaxValue)]
         public int RentalId { get; set; }
 
         public DateTime Start { get; set; }
